requirements.txt
langchain==0.1.10
ollama==0.1.5
faiss-cpu==1.7.4
chromadb==0.4.24
streamlit==1.31.1
mysql-connector-python==8.0.33
pandas==2.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests>=2.0